// src/api/main/model/githubComVeresusTlApiInternalModelOrderOut.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * TechLogist API
 * API сервер для управления логистикой
 * OpenAPI spec version: 1.0
 */
import type { GithubComVeresusTlApiInternalClientsTlOrdersClientDtoOrderOutStatus } from './githubComVeresusTlApiInternalClientsTlOrdersClientDtoOrderOutStatus';

/**
 * DTO для передачи информации о заявке
 */
export interface GithubComVeresusTlApiInternalModelOrderOut {
  cargo_description?: string;
  cargo_name?: string;
  cargo_type_name?: string;
  cargo_weight?: number;
  depart_loc_name?: string;
  dispatcher_name?: string;
  driver_name?: string;
  goal_loc_name?: string;
  id?: number;
  is_urgent?: boolean;
  order_status_name?: GithubComVeresusTlApiInternalClientsTlOrdersClientDtoOrderOutStatus;
  photo_id?: string;
  time?: string;
}